Update tests to replace mock http client with http test server, and move helper function tests to a seperate file
